+++ CHANGELOG.md
@@ -38,6 +38,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### New features
 
+* Git-format diff hunk headers now include nearby source symbols for many common
+  programming and markup languages.
+
 * New `merge_point()` revset function which (similar to `fork_point`) finds the
   point where multiple branches merge.